+++ .gitignore
@@ -129,3 +129,4 @@ dmypy.json
 .pyre/
 
 /demos
+/checkpoints
